Updated Contact.html
Added Abdullahs work to contact.html.
Edited the class name to match the class name in CSS (centerbox).
Edited the footer, gave it a unique ID for the contact page because it
would not stay on the bottom of the page without some adjusting. The
adjustments caused the footers on the index.html to move, so I gave the
contact page footer a unique ID.

I also added the webkit and mozikit box-shadow CSS.